refactor(graduate): 폴더 네이밍 스타일 통일
