+++ apps/graduate/src/pages/admin/filePreview/utils/index.ts
@@ -1,0 +1,2 @@
+export { default as transformCertificateResponse } from './transformCertificateFileResponse';
+export { default as transformThesisResponse } from './transformThesisFileResponse';
