+++ apps/graduate/src/pages/admin/schedule/constants/schedule.ts
