+++ apps/graduate/src/pages/admin/schedule/ui/ScheduleEditModal.tsx
@@ -8,7 +8,7 @@ import { modalStyles } from '~/shared/config';
 import { DATE_FORMAT } from '~/shared/constants';
 import { useToast } from '~/shared/hooks';
 
-import { SUBMISSION_TYPE_OPTIONS } from '../constant/schedule.ts';
+import { SUBMISSION_TYPE_OPTIONS } from '../constants';
 import { useUpdateSchedule } from '../hooks';
 import type { ScheduleItem } from '../model';
 import * as style from '../styles/ScheduleEditModal.css.ts';
